btservice/AdaperState: handle ENABLED_READY in OffState
If the adapter takes too long (more than ENABLE_TIMEOUT_DELAY) to enable,
the AdapterState will go into the OffState. In this state, ENABLED_READY
messages are ignored. When this happens, the bt vendor code thinks
it's enabled, but the AdapterState doesn't. Thus, the two layers will
stay out-of-sync until a reboot.

Handling the ENABLED_READY message while in OffState seems to be a
workable solution to the situation outlined above.

Change-Id: I6d090440a085b433e50dbb7edee20eb1efa510de